Phase 03- llm_response/requirements.txt
python-dotenv>=1.0.0
google-generativeai>=0.8.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pytest>=7.0.0
httpx>=0.25.0
